fix(send): remove old screens, optimise components with memo()
