+++ src/navigation/bottom-sheet/SendNavigation.tsx
@@ -1,4 +1,4 @@
-import React, { ReactElement, useMemo } from 'react';
+import React, { ReactElement, useMemo, memo } from 'react';
 import {
 	createNativeStackNavigator,
 	NativeStackNavigationProp,
@@ -7,14 +7,14 @@ import { TransitionPresets } from '@react-navigation/stack';
 import { useSelector } from 'react-redux';
 
 import BottomSheetWrapper from '../../components/BottomSheetWrapper';
-import AddressAndAmount from '../../screens/Wallets/SendOnChainTransaction2/AddressAndAmount';
-import FeeRate from '../../screens/Wallets/SendOnChainTransaction2/FeeRate';
-import FeeCustom from '../../screens/Wallets/SendOnChainTransaction2/FeeCustom';
-import Tags from '../../screens/Wallets/SendOnChainTransaction2/Tags';
-import ReviewAndSend from '../../screens/Wallets/SendOnChainTransaction2/ReviewAndSend';
-import SendAssetPickerList from '../../screens/Wallets/SendOnChainTransaction2/SendAssetPickerList';
-import Result from '../../screens/Wallets/SendOnChainTransaction2/Result';
-import CoinSelection from '../../screens/Wallets/SendOnChainTransaction2/CoinSelection';
+import AddressAndAmount from '../../screens/Wallets/SendOnChainTransaction/AddressAndAmount';
+import FeeRate from '../../screens/Wallets/SendOnChainTransaction/FeeRate';
+import FeeCustom from '../../screens/Wallets/SendOnChainTransaction/FeeCustom';
+import Tags from '../../screens/Wallets/SendOnChainTransaction/Tags';
+import ReviewAndSend from '../../screens/Wallets/SendOnChainTransaction/ReviewAndSend';
+import SendAssetPickerList from '../../screens/Wallets/SendOnChainTransaction/SendAssetPickerList';
+import Result from '../../screens/Wallets/SendOnChainTransaction/Result';
+import CoinSelection from '../../screens/Wallets/SendOnChainTransaction/CoinSelection';
 import AuthCheck from '../../components/AuthCheck';
 import { NavigationContainer } from '../../styles/components';
 import {
@@ -87,4 +87,4 @@ const SendNavigation = (): ReactElement => {
 	);
 };
 
-export default SendNavigation;
+export default memo(SendNavigation);
