+++ src/screens/Wallets/SendOnChainTransaction/AddressAndAmount.tsx
@@ -30,6 +30,7 @@ import NavigationHeader from '../../../components/NavigationHeader';
 import AmountToggle from '../../../components/AmountToggle';
 import Button from '../../../components/Button';
 import Tag from '../../../components/Tag';
+import OnChainNumberPad from '../SendOnChainTransaction/OnChainNumberPad';
 import Store from '../../../store/types';
 import { IOutput } from '../../../store/types/wallet';
 import { getTransactionOutputValue } from '../../../utils/wallet/transactions';
@@ -281,6 +282,7 @@ const AddressAndAmount = ({ index = 0, navigation }): ReactElement => {
 					/>
 				</View>
 			</View>
+			<OnChainNumberPad />
 		</ThemedView>
 	);
 };
