+++ src/screens/Wallets/SendOnChainTransaction/FeeCustom.tsx
@@ -9,6 +9,7 @@ import Button from '../../../components/Button';
 import { useTransactionDetails } from '../../../hooks/transaction';
 import { toggleView } from '../../../store/actions/user';
 import FeeCustomToggle from './FeeCustomToggle';
+import FeeNumberPad from './FeeNumberPad';
 
 const FeeRate = ({ navigation }): ReactElement => {
 	const insets = useSafeAreaInsets();
@@ -65,6 +66,7 @@ const FeeRate = ({ navigation }): ReactElement => {
 					/>
 				</View>
 			</View>
+			<FeeNumberPad />
 		</ThemedView>
 	);
 };
