+++ src/store/shapes/user.ts
@@ -19,8 +19,6 @@ export const defaultUserShape = {
 	backupVerified: false,
 	// Used to control various views throughout the app. (Modals, bottom-sheets, etc.)
 	viewController: {
-		send: { ...defaultViewController },
-		sendAssetPicker: { ...defaultViewController },
 		sendNavigation: { ...defaultViewController },
 		receiveNavigation: { ...defaultViewController },
 		backupPrompt: { ...defaultViewController },
@@ -32,8 +30,6 @@ export const defaultUserShape = {
 		boostPrompt: { ...defaultViewController },
 		activityTagsPrompt: { ...defaultViewController },
 		newTxPrompt: { ...defaultViewController },
-		coinSelection: { ...defaultViewController },
-		feePicker: { ...defaultViewController },
 		profileAddDataForm: { ...defaultViewController },
 		profileAddLinkForm: { ...defaultViewController },
 		addContactModal: { ...defaultViewController },
